move format to utils
